fixed packaging order
