--- src/main/java/animals/Chicken.java
@@ -1,4 +1,0 @@
-package animals;
-
-public class Chicken {
-}
