--- src/main/java/animals/Horse.java
@@ -1,4 +1,0 @@
-package animals;
-
-public class Horse {
-}
