--- src/main/java/calendar/Days.java
@@ -1,4 +1,0 @@
-package calendar;
-
-public enum Days {
-}
