--- src/main/java/calendar/Weeks.java
@@ -1,4 +1,0 @@
-package calendar;
-
-public class Weeks {
-}
