+++ src/main/java/com/zipcodewilmington/froilansfarm/animals/Chicken.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.animals;
+
+public class Chicken {
+}
