+++ src/main/java/com/zipcodewilmington/froilansfarm/calendar/Days.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.calendar;
+
+public enum Days {
+}
