+++ src/main/java/com/zipcodewilmington/froilansfarm/calendar/Weeks.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.calendar;
+
+public class Weeks {
+}
