+++ src/main/java/com/zipcodewilmington/froilansfarm/crops/PotatoPlants.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.crops;
+
+public class PotatoPlants {
+}
