+++ src/main/java/com/zipcodewilmington/froilansfarm/crops/WatermelonPlants.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.crops;
+
+public class WatermelonPlants {
+}
