+++ src/main/java/com/zipcodewilmington/froilansfarm/edibles/Watermelon.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.edibles;
+
+public class Watermelon {
+}
