+++ src/main/java/com/zipcodewilmington/froilansfarm/interfaces/Eater.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.interfaces;
+
+public interface Eater {
+}
