+++ src/main/java/com/zipcodewilmington/froilansfarm/interfaces/Produces.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.interfaces;
+
+public interface Produces {
+}
