+++ src/main/java/com/zipcodewilmington/froilansfarm/mainFarm/ChickeCoop.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.mainFarm;
+
+public class ChickeCoop {
+}
