+++ src/main/java/com/zipcodewilmington/froilansfarm/mainFarm/Farm.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.mainFarm;
+
+abstract class Farm {
+}
