+++ src/main/java/com/zipcodewilmington/froilansfarm/mainFarm/Stable.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.mainFarm;
+
+public class Stable {
+}
