+++ src/main/java/com/zipcodewilmington/froilansfarm/vehicles/CropDuster.java
@@ -1,0 +1,4 @@
+package com.zipcodewilmington.froilansfarm.vehicles;
+
+public class CropDuster {
+}
