--- src/main/java/crops/CornStalk.java
@@ -1,4 +1,0 @@
-package crops;
-
-public class CornStalk {
-}
