--- src/main/java/crops/GrapePlants.java
@@ -1,4 +1,0 @@
-package crops;
-
-public class GrapePlants {
-}
