--- src/main/java/crops/PotatoPlants.java
@@ -1,4 +1,0 @@
-package crops;
-
-public class PotatoPlants {
-}
