--- src/main/java/crops/TomatoPlants.java
@@ -1,4 +1,0 @@
-package crops;
-
-public class TomatoPlants {
-}
