--- src/main/java/crops/WatermelonPlants.java
@@ -1,4 +1,0 @@
-package crops;
-
-public class WatermelonPlants {
-}
