--- src/main/java/edibles/Corn.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Corn {
-}
