--- src/main/java/edibles/Eggs.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Eggs {
-}
