--- src/main/java/edibles/Grapes.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Grapes {
-}
