--- src/main/java/edibles/Potatos.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Potatos {
-}
