--- src/main/java/edibles/Tomatos.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Tomatos {
-}
