--- src/main/java/edibles/Watermelon.java
@@ -1,4 +1,0 @@
-package edibles;
-
-public class Watermelon {
-}
