--- src/main/java/interfaces/Eater.java
@@ -1,4 +1,0 @@
-package interfaces;
-
-public interface Eater {
-}
