--- src/main/java/interfaces/FarmVehicles.java
@@ -1,4 +1,0 @@
-package interfaces;
-
-public interface FarmVehicles {
-}
