--- src/main/java/interfaces/NoiseMaker.java
@@ -1,4 +1,0 @@
-package interfaces;
-
-public interface NoiseMaker {
-}
