--- src/main/java/interfaces/Produces.java
@@ -1,4 +1,0 @@
-package interfaces;
-
-public interface Produces {
-}
