--- src/main/java/interfaces/Rideable.java
@@ -1,4 +1,0 @@
-package interfaces;
-
-public interface Rideable {
-}
