--- src/main/java/mainFarm/ChickeCoop.java
@@ -1,4 +1,0 @@
-package mainFarm;
-
-public class ChickeCoop {
-}
