--- src/main/java/mainFarm/Farm.java
@@ -1,4 +1,0 @@
-package mainFarm;
-
-abstract class Farm {
-}
