--- src/main/java/mainFarm/FarmHouse.java
@@ -1,4 +1,0 @@
-package mainFarm;
-
-public class FarmHouse {
-}
