--- src/main/java/mainFarm/Field.java
@@ -1,4 +1,0 @@
-package mainFarm;
-
-public class Field {
-}
