--- src/main/java/mainFarm/Stable.java
@@ -1,4 +1,0 @@
-package mainFarm;
-
-public class Stable {
-}
