--- src/main/java/vehicles/Aircraft.java
@@ -1,4 +1,0 @@
-package vehicles;
-
-public class Aircraft {
-}
