--- src/main/java/vehicles/CropDuster.java
@@ -1,4 +1,0 @@
-package vehicles;
-
-public class CropDuster {
-}
