--- src/main/java/vehicles/Tractor.java
@@ -1,4 +1,0 @@
-package vehicles;
-
-public class Tractor {
-}
